pyproject.toml
[tool.poetry]
name = "overviewer-quotes-word-cloud"
version = "0.0.1"
description = ""
authors = ["ion"]

[tool.poetry.dependencies]
python = "^3.10"
wordcloud = "^1.8.2"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
